navigator bar is now the app-bar defined in header.tsx component. Then modifications to the home layout to improve ux.
